Universalized team_Kill_Compare function added save profile function
Made the targeted match stat an argument for the team_Stat_Compare function which allows for any stat from a match to be compared.

Also added a function that is used to save player profile info on a json file with the intent to minimize api calls by loading previously used summoner profiles.