Fixed issues with loading json data for profiles.json
Found that there was a memory bank issue with using an if statement to check if json file was empty. Now using a try/except statement to test if it is empty.

Currently working on properly iterating through the profiles in the json file to check if profile to be appended is a duplicate and needs to be skipped.